moved UmlApplication to its own files, and updated external links
